Move TextDisplayMediumAutoResize to own class
