+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/text/TextDisplayMedium.kt
@@ -1,15 +1,9 @@
 package app.k9mail.core.ui.compose.designsystem.atom.text
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import app.k9mail.core.ui.compose.theme2.MainTheme
 import androidx.compose.material3.Text as Material3Text
@@ -43,39 +37,5 @@ fun TextDisplayMedium(
         color = color,
         textAlign = textAlign,
         style = MainTheme.typography.displayMedium,
-    )
-}
-
-@Composable
-fun TextDisplayMediumAutoResize(
-    text: String,
-    modifier: Modifier = Modifier,
-    color: Color = Color.Unspecified,
-    textAlign: TextAlign? = null,
-) {
-    val style: TextStyle = MainTheme.typography.displayMedium
-    var shouldDraw by remember { mutableStateOf(false) }
-    var resizedTextStyle by remember { mutableStateOf(style) }
-
-    Material3Text(
-        text = text,
-        modifier = modifier.drawWithContent {
-            if (shouldDraw) {
-                drawContent()
-            }
-        },
-        color = color,
-        textAlign = textAlign,
-        softWrap = false,
-        style = resizedTextStyle,
-        onTextLayout = { result ->
-            if (result.didOverflowWidth) {
-                resizedTextStyle = resizedTextStyle.copy(
-                    fontSize = resizedTextStyle.fontSize * 0.95,
-                )
-            } else {
-                shouldDraw = true
-            }
-        },
     )
 }
